requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==4.2.16
# django-admin-shortcuts==3.0.1
django-adminlte3==0.1.6
django-cors-headers==4.4.0
django-filter==24.3
django-import-export==4.1.1
django-jet==1.0.8
django-semantic-admin==0.6.2
django-semantic-forms==0.1.6
djangocms-admin-style==3.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
pillow==10.4.0
PyJWT==2.9.0
sqlparse==0.5.1
tablib==3.5.0
